+++ packages/framework/src/hooks/__tests__/useEnsembleStorage.test.ts
@@ -27,4 +27,56 @@ describe("useEnsembleStorage", () => {
 
     expect(storage.get("test")).toBe("bar");
   });
+
+  test("overwrites existing keys", () => {
+    store.set(screenStorageAtom, {
+      test: {
+        foo: ["bar", "baz"],
+        dead: {
+          beef: "hello",
+        },
+      },
+    });
+
+    const { result } = renderHook(() => useEnsembleStorage());
+
+    const storage = result.current;
+
+    act(() => {
+      storage.set("test", {
+        foo: ["baz", "bar"],
+        beef: {
+          dead: "world",
+        },
+      });
+    });
+
+    expect(storage.get("test")).toMatchObject({
+      foo: ["baz", "bar"],
+      beef: {
+        dead: "world",
+      },
+    });
+  });
+
+  test("deletes keys", () => {
+    store.set(screenStorageAtom, {
+      test: {
+        foo: ["bar", "baz"],
+        dead: {
+          beef: "hello",
+        },
+      },
+    });
+
+    const { result } = renderHook(() => useEnsembleStorage());
+
+    const storage = result.current;
+
+    act(() => {
+      storage.delete("test");
+    });
+
+    expect(storage.get("test")).toBeUndefined();
+  });
 });