+++ packages/framework/src/state/screen.ts
@@ -1,6 +1,6 @@
 import { atom } from "jotai";
 import { focusAtom } from "jotai-optics";
-import { isObject, mergeWith } from "lodash-es";
+import { assign } from "lodash-es";
 import { type Response, type WebSocketConnection } from "../data";
 import type { EnsembleAppModel, EnsembleScreenModel } from "../shared";
 import type { WidgetState } from "./widget";
@@ -9,7 +9,7 @@ export interface ScreenContextDefinition {
   app?: EnsembleAppModel;
   model?: EnsembleScreenModel;
   data: ScreenContextData;
-  widgets: { [key: string]: WidgetState };
+  widgets: { [key: string]: WidgetState | undefined };
   inputs?: { [key: string]: unknown };
   [key: string]: unknown;
 }
@@ -40,11 +40,7 @@ export const screenDataAtom = atom(
   (get) => get(screenDataFocusAtom),
   (get, set, update) => {
     const currentData = get(screenDataFocusAtom);
-    const nextData = mergeWith({}, currentData, update, (objVal, srcVal) => {
-      if (isObject(objVal) && isObject(srcVal)) {
-        return srcVal;
-      }
-    });
+    const nextData = assign({}, currentData, update);
     set(screenDataFocusAtom, nextData);
   },
 );