+++ apps/starter/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @ensembleui/react-starter
 
+## 0.4.26
+
+### Patch Changes
+
+- Updated dependencies [549a952]
+- Updated dependencies [9eab1dc]
+  - @ensembleui/react-framework@0.2.2
+  - @ensembleui/react-runtime@0.2.2
+
+## 0.4.25
+
+### Patch Changes
+
+- Updated dependencies [071fe2c]
+- Updated dependencies [85f332a]
+- Updated dependencies [5163b57]
+- Updated dependencies [4e4d9d6]
+  - @ensembleui/react-runtime@0.2.1
+  - @ensembleui/react-framework@0.2.1
+
 ## 0.4.24
 
 ### Patch Changes